Move common functionality between sessions to Session trait.
