+++ src-tauri/src/app/state/mod.rs
@@ -12,13 +12,15 @@ pub mod analysis;
 /// Declares state objects that are unique to the sketchbook editor window.
 pub mod editor;
 
+use crate::app::state::_undo_stack::UndoStack;
+use crate::debug;
 pub use _consumed::Consumed;
 pub use _state_app::AppState;
 pub use _state_atomic::AtomicState;
 //pub use _state_map::MapState;
 
 pub type DynSessionState = Box<(dyn SessionState + Send + 'static)>;
-pub type DynSession = Box<(dyn Session + Send + 'static)>;
+pub type DynSession = Box<(dyn StackSession + Send + 'static)>;
 
 pub trait SessionState {
     /// Modify the session state using the provided `event`. The possible outcomes are
@@ -108,11 +110,178 @@ pub trait SessionHelper {
     }
 }
 
-pub trait Session: SessionState {
+/// A Session with a [UndoStack] with events.
+///
+/// Sessions perform user actions, or communicate with different sessions via messages.
+pub trait StackSession: SessionState {
     /// Perform a user action on this session state object. This usually involves propagating
     /// the events to the internal [SessionState] objects and collecting the results into a
     /// single [StateChange] entry.
-    fn perform_action(&mut self, action: &UserAction) -> Result<StateChange, DynError>;
+    ///
+    /// In this top-level method, we explicitly test for undo-stack actions. Once that is done,
+    /// the processing continues via [perform_categorized_action].
+    fn perform_action(&mut self, action: &UserAction) -> Result<StateChange, DynError> {
+        // Explicit test for undo-stack actions.
+        // TODO:
+        //  Figure out a nicer way to do this. Probably modify the `Consumed` enum?
+        //  We basically need a way to say "restart with these events, but as an
+        //  Irreversible action that won't reset the stack."
+        'undo: {
+            if action.events.len() == 1 {
+                let event = &action.events[0];
+                if event.path.len() == 2 && event.path[0] == "undo_stack" {
+                    let action = match event.path[1].as_str() {
+                        "undo" => {
+                            let Some(undo) = self.undo_stack_mut().undo_action() else {
+                                return AeonError::throw("Nothing to undo.");
+                            };
+                            undo
+                        }
+                        "redo" => {
+                            let Some(redo) = self.undo_stack_mut().redo_action() else {
+                                return AeonError::throw("Nothing to redo.");
+                            };
+                            redo
+                        }
+                        _ => break 'undo,
+                    };
+                    let mut state_change = self.perform_categorized_action(&action, true)?;
+                    self.append_stack_updates(&mut state_change.events);
+                    return Ok(state_change);
+                }
+            }
+        }
+        self.perform_categorized_action(action, false)
+    }
+
+    /// Perform a user action on this session state object, with additional information whether
+    /// the action should bypass the undo-redo stack.
+    ///
+    /// This method assumes the action was already categorized into one of `undo` (stack should be
+    /// bypassed) or `regular` (goes to the undo stack).
+    /// If you want to run the full process including categorizing the action, use [perform_action].
+    fn perform_categorized_action(
+        &mut self,
+        action: &UserAction,
+        ignore_stack: bool,
+    ) -> Result<StateChange, DynError> {
+        // Events that need to be consume (last to first) in order to complete this action.
+        let mut to_perform = action.events.clone();
+        to_perform.reverse();
+
+        // The events representing successful state changes.
+        let mut state_changes: Vec<Event> = Vec::new();
+        // The events that can be used to create a redo stack entry if the action is reversible.
+        let mut reverse: Option<Vec<(Event, Event)>> =
+            if ignore_stack { None } else { Some(Vec::new()) };
+        let mut reset_stack = false;
+
+        while let Some(event) = to_perform.pop() {
+            let event_path = event.path.iter().map(|it| it.as_str()).collect::<Vec<_>>();
+            debug!(
+                "Executing event to session {}: `{:?}`.",
+                self.id(),
+                event_path
+            );
+            let result = match self.perform_event(&event, &event_path) {
+                Ok(result) => result,
+                Err(error) => {
+                    // TODO:
+                    //  We should probably first emit the state change and then the
+                    //  error, because now we are losing state of compound actions that fail.
+                    return Err(error);
+                }
+            };
+            match result {
+                Consumed::Reversible {
+                    state_change,
+                    perform_reverse,
+                } => {
+                    state_changes.push(state_change);
+                    if let Some(reverse) = reverse.as_mut() {
+                        // If we can reverse this action, save the events.
+                        reverse.push(perform_reverse);
+                    }
+                }
+                Consumed::Irreversible {
+                    state_change,
+                    reset,
+                } => {
+                    state_changes.push(state_change);
+                    if reset {
+                        // We cannot reverse this event, but the rest can be reversed.
+                        reverse = None;
+                        reset_stack = true;
+                    }
+                }
+                Consumed::Restart(mut events) => {
+                    // Just push the new events to the execution stack and continue
+                    // to the next event.
+                    events.reverse();
+                    while let Some(e) = events.pop() {
+                        to_perform.push(e);
+                    }
+                }
+                Consumed::InputError(error) => {
+                    // TODO:
+                    //  The same as above. We should report this as a separate event from the
+                    //  state change that was performed.
+                    return Err(error);
+                }
+                Consumed::NoChange => {
+                    // Do nothing.
+                }
+            }
+        }
+        // If the action is not irreversible, we should add an entry to the undo stack.
+        if let Some(events) = reverse {
+            if !events.is_empty() {
+                // Only add undo action if the stack is not empty.
+                let mut perform = Vec::new();
+                let mut reverse = Vec::new();
+                for (p, r) in events {
+                    perform.push(p);
+                    reverse.push(r);
+                }
+                // Obviously, the "reverse" events need to be execute in the opposite order
+                // compared to the "perform" events.
+                reverse.reverse();
+                let perform = UserAction { events: perform };
+                let reverse = UserAction { events: reverse };
+                if !self.undo_stack_mut().do_action(perform, reverse) {
+                    // TODO: Not match we can do here, maybe except issuing a warning.
+                    self.undo_stack_mut().clear();
+                }
+
+                // Notify about the changes in the stack state.
+                // TODO: Maybe we don't need to emit this always.
+                self.append_stack_updates(&mut state_changes);
+            }
+        } else if !ignore_stack && reset_stack {
+            debug!(
+                "Back stack (of session {}) cleared due to irreversible action.",
+                self.id()
+            );
+            self.undo_stack_mut().clear();
+        }
+
+        Ok(StateChange {
+            events: state_changes,
+        })
+    }
+
+    fn append_stack_updates(&self, state_changes: &mut Vec<Event>) {
+        let can_undo = serde_json::to_string(&self.undo_stack().can_undo());
+        let can_redo = serde_json::to_string(&self.undo_stack().can_redo());
+        state_changes.push(Event::build(
+            &["undo_stack", "can_undo"],
+            Some(can_undo.unwrap().as_str()),
+        ));
+        state_changes.push(Event::build(
+            &["undo_stack", "can_redo"],
+            Some(can_redo.unwrap().as_str()),
+        ));
+    }
 
     /// Process a message sent to this session state object.
     ///
@@ -128,4 +297,10 @@ pub trait Session: SessionState {
     /// Returns the string identifier of this particular session. Each session identifier must
     /// be unique within the application.
     fn id(&self) -> &str;
+
+    /// Returns an immutable reference to session's undo stack.
+    fn undo_stack(&self) -> &UndoStack;
+
+    /// Returns a mutable reference to session's undo stack.
+    fn undo_stack_mut(&mut self) -> &mut UndoStack;
 }
